refactor(store): move ProductDetailContainer and types to correct directories
