--- src/store/types.ts
@@ -1,5 +1,0 @@
-export interface Category {
-  id: string
-  label: string
-  count: number
-}
